// src/app/model/person.ts
export class Person {
    ANNUAL_INCOME?: number;
    BRONZE?: number;
    DOB?: string;
    EMPLOYMENT_STATUS?: string;
    GOLD?: number;
    HEIGHT?: number;
    MARITAL_STATUS?: string;
    OPTIONAL_INSURED?: number;
    PEOPLE_COVERED?: number;
    PLATINUM?: number;
    PURCHASED?: string;
    SILVER?: number;
    TOBACCO?: string;
    WEIGHT?: number;
    address?: string;
    city?: string;
    collection_id?: string;
    email?: string;
    id?: string;
    latitude?: number;
    longitude?: number;
    name?: string;
    sex?: string;
    state?: string;
}